Refactor C-weighting filter design to core analysis (Fix Imports)
Extract `_design_c_weighting` from `src/gui/widgets/settings.py` to `src/core/analysis.py` as `design_c_weighting`.
This submission resolves a blocking `ImportError` by correctly handling the existing imports in `src/core/analysis.py`.

* `src/core/analysis.py`: Added `design_c_weighting(sr)` and verified `scipy.signal` and `numpy` imports exist.
* `src/gui/widgets/settings.py`: Replaced local `_design_c_weighting` with import.
* `tests/logic_verification/analysis/test_c_weighting_design.py`: Added comprehensive logic verification tests.
* `tests/logic_verification/test_settings_c_weighting.py`: Deleted obsolete test.
