+++ src/core/fft_manager.py
@@ -1,4 +1,3 @@
-import functools
 import numpy as np
 import multiprocessing
 import threading
@@ -311,7 +310,6 @@ class FFTManager:
 
             # Use MEASURE for warmup to ensure peak performance
             self.get_plan(size, "float64", flags=("FFTW_MEASURE",))
-            self.get_plan(size, "float32", flags=("FFTW_MEASURE",))
 
         # Save wisdom at the end of warmup to capture any new measurements
         if callback:
@@ -346,20 +344,6 @@ class FFTManager:
             "tukey",
             "taylor",
         ]
-
-
-@functools.lru_cache(maxsize=16)
-def get_dpss_windows(N, NW=3, Kmax=None):
-    """
-    Get DPSS windows, caching them for performance.
-    """
-    # Lazy import to avoid hard dependency at module level
-    from scipy.signal.windows import dpss
-
-    if Kmax is None:
-        Kmax = int(2 * NW - 1)
-
-    return dpss(N, NW, int(Kmax))
 
 
 # Global instance for easy access
